Used aws service

+++ PasteBinClone/PasteBinClone.Application/Services/PasteService.cs
@@ -12,55 +12,71 @@ using System.Threading.Tasks;
 namespace PasteBinClone.Application.Services
 {
     public class PasteService(IPasteRepository pasteRepository, 
-        IMapper mapper) : IPasteService
+        IMapper mapper,
+        IAmazonStorageService amazonStorage) : IPasteService
     {
         private readonly IPasteRepository _pasteRepository = pasteRepository;
         private readonly IMapper _mapper = mapper;
+        private readonly IAmazonStorageService _amazonStorage = amazonStorage;
 
         public async Task<bool> CreatePaste(PasteDto pasteDto)
         {
-            Paste paste = new()
-            {
-                Title = pasteDto.Title,
-                BodyUrl = "",
-                IsPublic = false,
-                LifeTime = DateTime.Now,
-                CategoryId = 0,
-                TypeId = 0,
-                LanguageId = 0,
-                UserId = ""
-            };
+            (bool isCreate, string pasteId ) = 
+                await _amazonStorage.UploadFile(pasteDto.Body);
 
-            //if the creation was successful, the method will return true
-            bool result = await _pasteRepository.Create(paste);
+            if (isCreate && pasteId != null)
+            {
+                Paste paste = new()
+                {
+                    Title = pasteDto.Title,
+                    BodyUrl = pasteId,
+                    IsPublic = pasteDto.IsPublic,
+                    CreateAt = DateTime.Now,
+                    ExpireAt = pasteDto.ExpireAt,
+                    CategoryId = pasteDto.CategoryId,
+                    TypeId = pasteDto.TypeId,
+                    LanguageId = pasteDto.LanguageId,
+                    UserId = pasteDto.UserId
+                };
 
-            if (result)
-            {
-                Log.Information("Object {@i} created successfully.", paste.Id);
-                return true;    
+                //if the creation was successful, the method will return true
+                bool result = await _pasteRepository.Create(paste);
+
+                if (result)
+                {
+                    Log.Information("Object {@i} created successfully.", paste.Id);
+                    return true;
+                }
+                else
+                {
+                    Log.Error("Error creating object.");
+                    return false;
+                }
             }
-            else
-            {
-                Log.Error("Error creating object.");
-                return false;
-            }
+
+            return false;
         }
 
         public async Task<bool> DeletePaste(Guid id)
         {
-            //if the deletion was successful, the method will return true
-            bool result = await _pasteRepository.Delete(id);
+            Paste paste = await _pasteRepository.GetById(id);
 
-            if (result)
+            bool awsDeletionResult = await _amazonStorage.DeleteFile(paste.BodyUrl);
+
+            //if the deletion was successful, the method will return true
+            if (awsDeletionResult)
             {
-                Log.Information("Object {@i} successfully deleted.", id);
-                return true;
+                bool result = await _pasteRepository.Delete(id);
+
+                if (result)
+                {
+                    Log.Information("Object {@i} successfully deleted.", id);
+                    return true;
+                }
             }
-            else
-            {
-                Log.Error("Object deletion error.");
-                return false;
-            }
+
+            Log.Error("Object deletion error.");
+            return false;
         }
 
         public async Task<IEnumerable<PasteDto>> GetAllPaste()
@@ -86,20 +102,31 @@ namespace PasteBinClone.Application.Services
                 return null;
             }
 
-            return _mapper.Map<PasteDto>(paste);
+            string pasteBody = await _amazonStorage.GetFile(paste.BodyUrl);
+
+            var pasteDto = _mapper.Map<PasteDto>(paste);
+            pasteDto.Body = pasteBody;
+
+            return pasteDto;
         }
 
         public async Task<bool> UpdatePaste(PasteDto pasteDto)
         {
             Paste paste = _mapper.Map<Paste>(pasteDto);
 
-            //if the update was successful, the method will return true
-            bool result = await _pasteRepository.Update(paste);
+            bool updateResult = await _amazonStorage
+                .UpdateFile(pasteDto.BodyUrl, pasteDto.Body);
 
-            if (result)
+            //if the update was successful, the method will return true
+            if (updateResult)
             {
-                Log.Information("Object {@i} updated.", pasteDto.Id);
-                return true;
+                bool result = await _pasteRepository.Update(paste);
+
+                if (result)
+                {
+                    Log.Information("Object {@i} updated.", pasteDto.Id);
+                    return true;
+                }
             }
 
             Log.Error("Object update error.");
